# user_swagger_based_api/api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: |
    This is a User service API
  version: 1.0.0
  title: Swagger User
tags:
- name: user
  description: Operations about user
  externalDocs:
    url: http://swagger.io
# schemes:
# - http
paths:
  /users:
    post:
      tags:
      - user
      summary: Create or update user
      description: This can only be done by the logged in user.
      operationId: createUserOrUpdate
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Created user object
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        default:
          description: successful operation
    get:
      tags:
      - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: getAllUsers
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Get all users
        required: true
        schema:
          type: array
          items:
            $ref: '#/definitions/User'
      responses:
        default:
          description: successful operation
  /users/{id}:
    get:
      tags:
      - user
      summary: Get user by user id
      operationId: getUserById
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: The id that needs to be fetched. Use user1 for testing. 
        required: true
        type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        400:
          description: Invalid login supplied
        404:
          description: User not found
    delete:
      tags:
      - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: The name that needs to be deleted
        required: true
        type: string
      responses:
        404:
          description: User not found
definitions:
  User:
    type: object
    properties:
      id:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
    xml:
      name: User
# Added by API Auto Mocking Plugin
host: localhost
basePath: /YaroslavKratt/nodejs_mentoring/1.0.0
schemes:
 - http